Bug 13036 Boost.Regex: overflow in max_state_count
This patch partially addresses 13036. It partially addresses this issue
by using an unsigned integer overflow check on a signed integer value
(the states variable).

A signed overflow check works because the states variable is initialized
with a value greater than or equal to zero and the CERT solution for
checking overflow on signed values when both operands are positive and
the same simplifies to the unsigned overflow check.

This patch partially addresses 13036 because the declaration of the
states variable uses ptrdiff_t, but states is initialized using a value
declared as size_t. A guarantee that states does not overflow on
initialization needs to be established. This guarantee may already exist
in the code or there may be another bug.